이력서 평가 로직 1차 구현

+++ src/main/java/com/example/resume/dto/EvaluationResponseDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 public record EvaluationResponseDto(
         Long id,
-        Integer score,
+        Double score,
         LocalDateTime evaluatedAt,
         String comment
 ) {
